rename and move types
