+++ src/types/image/index.ts
@@ -1,9 +1,9 @@
-export interface Dimensions {
-    width: number;
-    height: number;
-}
-
 export interface BlobOptions {
     type?: string;
     quality?: number;
+}
+
+export interface Dimensions {
+    width: number;
+    height: number;
 }
