+++ src/types/upload/index.ts
@@ -36,6 +36,11 @@ export interface GeoLocation {
     longitude: number;
 }
 
+export interface Dimensions {
+    width: number;
+    height: number;
+}
+
 export interface ParsedMetadataJSON {
     creationTime: number;
     modificationTime: number;
