+++ src/utils/image/index.ts
@@ -1,6 +1,6 @@
 // these utils only work in env where OffscreenCanvas is available
 
-import { BlobOptions, Dimensions } from 'types/file/browserFile';
+import { BlobOptions, Dimensions } from 'types/image';
 import { enlargeBox } from 'utils/machineLearning';
 import { Box } from '../../../thirdparty/face-api/classes';
